[Refactor] #131 - 점주 발주 이력 테이블/상세 모달 컴포넌트 분리
- StoreOrderHistoryTable 컴포넌트 생성
- StoreOrderDetailModal 컴포넌트 생성
- StoreOrderManageView에서 인라인 코드 제거 후 컴포넌트로 교체
