+++ frontend/src/views/store/StoreOrderManageView.vue
@@ -182,129 +182,10 @@
     </div>
 
 
-    <div v-if="activeTab === 'history'">
-      <div class="bg-white border border-gray-200 rounded-lg overflow-hidden">
-        <table class="w-full text-sm text-left">
-          <thead>
-            <tr class="border-b border-gray-200 bg-gray-50">
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">품목</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">처리</th>
-            </tr>
-          </thead>
-          <tbody class="divide-y divide-gray-100">
-            <tr v-for="h in orderHistory" :key="h.idx"
-              class="hover:bg-gray-50/50 transition-colors cursor-pointer"
-              @click="openHistoryDetail(h)">
-              <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ h.idx }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-bold px-2 py-0.5 rounded"
-                  :class="h.ordersType === 'AUTO' ? 'bg-blue-50 text-blue-600 border border-blue-200' : 'bg-purple-50 text-purple-600 border border-purple-200'">
-                  {{ ORDER_TYPE_LABEL[h.ordersType] }}
-                </span>
-              </td>
-              <td class="px-5 py-3.5 font-semibold text-gray-900">
-                {{ h.ordersItemList[0]?.productName }}
-                <span v-if="h.ordersItemList.length > 1" class="text-xs text-gray-400 font-normal"> 외 {{ h.ordersItemList.length - 1 }}건</span>
-              </td>
-              <td class="px-5 py-3.5 font-semibold text-gray-700">
-                ₩ {{ h.price?.toLocaleString() }}
-              </td>
-              <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ h.createdAt?.replace('T', ' ').slice(0, 16) }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-bold px-2 py-0.5 rounded" :class="historyStatusCls(h.ordersStatus)">
-                  {{ ORDER_STATUS_LABEL[h.ordersStatus] }}
-                </span>
-              </td>
-              <td class="px-5 py-3.5">
-                <button v-if="h.ordersStatus === 'CONFIRMED'"
-                  @click.stop="cancelOrder(h)"
-                  class="px-3 py-1.5 text-xs font-semibold rounded-lg border border-red-200 text-red-500 bg-red-50 hover:bg-red-500 hover:text-white hover:cursor-pointer transition-colors">
-                  취소
-                </button>
-              </td>
-            </tr>
-            <tr v-if="orderHistory.length === 0">
-              <td colspan="7" class="px-5 py-10 text-center text-sm text-gray-400">발주 이력이 없습니다.</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
+    <StoreOrderHistoryTable v-if="activeTab === 'history'" :orders="orderHistory"
+      @open-detail="openHistoryDetail" @cancel="cancelOrder" />
 
-    <!-- 이력 상세 모달 -->
-    <div v-if="showHistoryDetail" class="fixed inset-0 z-50 flex items-center justify-center">
-      <div class="absolute inset-0 bg-black/40" @click="showHistoryDetail = false"></div>
-      <div class="relative bg-white rounded-xl w-full max-w-md border border-gray-200 shadow-xl">
-        <div class="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
-          <div>
-            <h3 class="font-bold text-gray-900">발주 상세</h3>
-            <p class="text-xs text-gray-400 font-mono mt-0.5">No. {{ selectedHistory?.idx }}</p>
-          </div>
-          <button @click="showHistoryDetail = false" class="text-gray-400 hover:text-gray-600 cursor-pointer">✕</button>
-        </div>
-        <div v-if="selectedHistory" class="p-6 space-y-4 text-sm">
-          <div class="grid grid-cols-2 gap-4">
-            <div>
-              <p class="text-xs text-gray-400 mb-1">유형</p>
-              <span class="text-xs font-bold px-2 py-0.5 rounded border"
-                :class="selectedHistory.ordersType === 'AUTO' ? 'bg-blue-50 text-blue-600 border-blue-200' : 'bg-purple-50 text-purple-600 border-purple-200'">
-                {{ ORDER_TYPE_LABEL[selectedHistory.ordersType] }}
-              </span>
-            </div>
-            <div>
-              <p class="text-xs text-gray-400 mb-1">상태</p>
-              <span class="text-xs font-bold px-2 py-0.5 rounded" :class="historyStatusCls(selectedHistory.ordersStatus)">
-                {{ ORDER_STATUS_LABEL[selectedHistory.ordersStatus] }}
-              </span>
-            </div>
-            <div class="col-span-2">
-              <p class="text-xs text-gray-400 mb-1">발주일시</p>
-              <p class="text-gray-700 font-mono text-xs">{{ selectedHistory.createdAt?.replace('T', ' ').slice(0, 16) }}</p>
-            </div>
-          </div>
-          <div>
-            <p class="text-xs text-gray-400 mb-2">품목 목록</p>
-            <div class="border border-gray-100 rounded-lg overflow-hidden">
-              <table class="w-full text-sm">
-                <thead class="bg-gray-50">
-                  <tr>
-                    <th class="px-4 py-2.5 text-left text-[10px] font-bold text-gray-400 uppercase">품목명</th>
-                    <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">수량</th>
-                    <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">단가</th>
-                    <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">금액</th>
-                  </tr>
-                </thead>
-                <tbody class="divide-y divide-gray-100">
-                  <tr v-for="item in selectedHistory.ordersItemList" :key="item.idx">
-                    <td class="px-4 py-2.5 text-gray-800 font-semibold">{{ item.productName }}</td>
-                    <td class="px-4 py-2.5 text-right text-gray-600">{{ item.count.toLocaleString() }}개</td>
-                    <td class="px-4 py-2.5 text-right text-xs text-gray-500">₩ {{ item.unitPrice.toLocaleString() }}</td>
-                    <td class="px-4 py-2.5 text-right font-bold text-blue-600">₩ {{ (item.unitPrice * item.count).toLocaleString() }}</td>
-                  </tr>
-                </tbody>
-                <tfoot class="bg-gray-50 border-t border-gray-200">
-                  <tr>
-                    <td colspan="3" class="px-4 py-2.5 text-right text-xs font-bold text-gray-500">합계</td>
-                    <td class="px-4 py-2.5 text-right font-black text-blue-600">
-                      ₩ {{ selectedHistory.price?.toLocaleString() }}
-                    </td>
-                  </tr>
-                </tfoot>
-              </table>
-            </div>
-          </div>
-        </div>
-        <div class="px-6 py-4 border-t border-gray-100 flex justify-end">
-          <button @click="showHistoryDetail = false"
-            class="px-4 py-2 text-sm font-semibold text-gray-600 border border-gray-200 rounded-lg hover:bg-gray-50 cursor-pointer">닫기</button>
-        </div>
-      </div>
-    </div>
+    <StoreOrderDetailModal :order="selectedHistory" @close="selectedHistory = null" />
 
     <div v-if="isModalOpen" class="fixed inset-0 z-50 flex items-center justify-center bg-black/60 backdrop-blur-[2px]">
       <div class="bg-white rounded-xl shadow-2xl w-full max-w-md overflow-hidden animate-modal-up">
@@ -367,6 +248,8 @@
 import { ref, computed, reactive, onMounted } from 'vue'
 import { ClipboardList, CreditCard, Sparkles, ChevronDown, Plus } from 'lucide-vue-next'
 import StoreManualOrderModal from '@/components/orders/StoreManualOrderModal.vue'
+import StoreOrderHistoryTable from '@/components/orders/StoreOrderHistoryTable.vue'
+import StoreOrderDetailModal from '@/components/orders/StoreOrderDetailModal.vue'
 import ordersApi from '@/api/orders'
 
 const PRODUCT_UNIT = {
@@ -400,18 +283,6 @@ const PRODUCT_PRICES = {
   '간장':       4000,
   '양파':       1500,
   '생수':       8000,
-}
-
-const ORDER_STATUS_LABEL = {
-  CONFIRMED: '확정',
-  APPROVE: '승인',
-  REJECT: '거절',
-  CANCELLED: '취소',
-}
-
-const ORDER_TYPE_LABEL = {
-  AUTO: '자동',
-  MANUAL: '수동',
 }
 
 const activeTab = ref('pending')
@@ -483,21 +354,11 @@ onMounted(() => {
   fetchOrderHistory()
 })
 
-const showHistoryDetail = ref(false)
 const selectedHistory = ref(null)
 
 function openHistoryDetail(h) {
   selectedHistory.value = h
-  showHistoryDetail.value = true
 }
-
-const HISTORY_STATUS_CLS = {
-  CONFIRMED:  'bg-green-50 text-green-700 border border-green-200',
-  APPROVE:    'bg-blue-50 text-blue-600 border border-blue-200',
-  REJECT:     'bg-red-50 text-red-600 border border-red-200',
-  CANCELLED:  'bg-red-50 text-red-500 border border-red-200',
-}
-const historyStatusCls = s => HISTORY_STATUS_CLS[s] ?? 'bg-gray-100 text-gray-500 border border-gray-200'
 
 const tabs = computed(() => [
   { id: 'pending', label: '제안 발주서', count: pendingOrders.value.length },
